Add scrollbars
Add scrollbars to various UI areas to be more friendly to low resolution
monitors. This works well down to vertical resolutions of 960. Below
that, things will get scrunched.

+++ Source/Fractorium/FinalRenderEmberController.h
@@ -38,7 +38,7 @@ struct FinalRenderGuiState
 	unsigned int m_ThreadCount;
 	unsigned int m_Width;
 	unsigned int m_Height;
-	unsigned int m_Quality;
+	double m_Quality;
 	unsigned int m_TemporalSamples;
 	unsigned int m_Supersample;
 };
